Move monthly dataset site extraction function to better module
